+++ packages/transaction-pay-controller/src/strategy/fiat/fiat-quotes.ts
@@ -42,15 +42,18 @@ const log = createModuleLogger(projectLogger, 'fiat-strategy');
 export async function getFiatQuotes(
   request: PayStrategyGetQuotesRequest,
 ): Promise<TransactionPayQuote<FiatQuote>[]> {
-  const { accountSupports7702, fiatPaymentMethod, messenger, transaction } =
-    request;
+  const {
+    accountSupports7702,
+    fiatPaymentMethod,
+    from: walletAddress,
+    messenger,
+    transaction,
+  } = request;
   const transactionId = transaction.id;
 
   const state = messenger.call('TransactionPayController:getState');
   const transactionData = state.transactionData[transactionId];
   const amountFiat = transactionData?.fiatPayment?.amountFiat;
-  const walletAddress = (transactionData?.accountOverride ??
-    transaction.txParams.from) as Hex;
   const requiredTokens = getRequiredTokens(transactionData?.tokens);
   const fiatAsset = deriveFiatAssetForFiatPayment(transaction, messenger);
 
@@ -81,6 +84,7 @@ export async function getFiatQuotes(
 
     const relayQuotes = await getRelayQuotes({
       accountSupports7702,
+      from: walletAddress,
       messenger,
       requests: [relayRequest],
       transaction,