+++ packages/transaction-pay-controller/src/strategy/fiat/fiat-submit-simple.ts
@@ -49,6 +49,7 @@ export async function submitSimpleRelay({
 
   const relayQuotes = await getRelayQuotes({
     accountSupports7702: request.accountSupports7702,
+    from: baseRequest.from,
     messenger,
     requests: [relayRequest],
     transaction,